% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneData}
\alias{geneData}
\title{Example temporal gene expression data from Schlamp et al. 2021}
\format{
A data frame with 1735 rows and 17 columns. Row names are the gene names and column names are the time point names.
}
\usage{
geneData
}
\description{
A data set containing gene expression measurements for 1735 genes at 17 time points.
}
\keyword{datasets}
