% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{priorMatrix}
\alias{priorMatrix}
\title{Example prior information/adjacency matrix for example dataset}
\format{
A symmetric matrix (data frame) with 1735 rows and 1735 columns. Row and column names are the gene names.
}
\usage{
priorMatrix
}
\description{
A symmetric "adjacency" matrix for the example gene expression dataset ('geneData'), as described in Section 3 of Venkatraman et al. Entry (i,j) is 1 if there is prior evidence of association between the i-th and j-th genes, 0 if the association is unlikely, and NA if the association is unknown.
}
\keyword{datasets}
